move test util code out of test module
